+++ Sources/StreamChatUI/ChatThread/ChatThreadVC.swift
@@ -276,7 +276,12 @@ open class ChatThreadVC: _ViewController,
         footerViewForMessage message: ChatMessage,
         at indexPath: IndexPath
     ) -> ChatMessageDecorationView? {
-        nil
+        guard components.threadRepliesCounterEnabled, message == messages.last, message.replyCount > 0 else {
+            return nil
+        }
+        let repliesCounterDecorationView = components.threadRepliesCounterDecorationView.init()
+        repliesCounterDecorationView.content = L10n.Message.Thread.Replies.count(message.replyCount)
+        return repliesCounterDecorationView
     }
 
     // MARK: - ChatMessageControllerDelegate