+++ Sources/StreamChatUI/Utils/Extensions/UIView+Extensions.swift
@@ -7,8 +7,7 @@ import UIKit
 extension UIView {
     // MARK: - `embed` family of helpers
 
-    @objc
-    func embed(_ subview: UIView, insets: NSDirectionalEdgeInsets = .zero) {
+    @objc func embed(_ subview: UIView, insets: NSDirectionalEdgeInsets = .zero) {
         addSubview(subview)
 
         NSLayoutConstraint.activate([